Moved large portions of core spefic code to their own partials
