+++ src/PerformanceTests/Common/Common.Data.projitems
@@ -13,7 +13,8 @@
     <Compile Include="$(MSBuildThisFileDirectory)AuditProfile.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)CheckGCServer.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)CheckPlatform.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Scenarios\BaseRunner.IConfigurationSource.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Scenarios\BaseRunner.v5.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Scenarios\BaseRunner.v6.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Scenarios\FakeTransport.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Shortcut\ShortcutBehavior.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Shortcut\ShortcutFeature.cs" />
