+++ src/PerformanceTests/Common/Scenarios/BaseRunner.cs
@@ -3,16 +3,8 @@ using Configuration = NServiceBus.EndpointConfiguration;
 #else
 using Configuration = NServiceBus.BusConfiguration;
 #endif
-#if Version7
-using NServiceBus.Configuration.AdvancedExtensibility;
-#endif
-#if Version6
-using NServiceBus.Configuration.AdvanceExtensibility;
-#endif
 using System;
 using System.Linq;
-using System.Reflection;
-using NServiceBus;
 using NServiceBus.Logging;
 using Tests.Permutations;
 using System.Collections.Generic;
@@ -149,180 +141,6 @@ public abstract partial class BaseRunner : IContext
             await Session.CloseWithSuppress().ConfigureAwait(false);
         }
     }
-
-#if Version5
-    async Task CreateOrPurgeAndDrainQueues()
-    {
-        var configuration = CreateConfiguration();
-        if (IsPurgingSupported) configuration.PurgeOnStartup(true);
-        ShortcutBehavior.Shortcut = true; // Required, as instance already receives messages before DrainMessages() is called!
-        var instance = Bus.Create(configuration).Start();
-        await DrainMessages().ConfigureAwait(false);
-        await new Session(instance).CloseWithSuppress().ConfigureAwait(false);
-    }
-
-    Task CreateSendOnlyEndpoint()
-    {
-        var configuration = CreateConfiguration();
-        var instance = Bus.CreateSendOnly(configuration);
-        Session = new Session(instance);
-        return Task.FromResult(0);
-    }
-
-    Task CreateEndpoint()
-    {
-        var configuration = CreateConfiguration();
-        configuration.CustomConfigurationSource(this);
-        configuration.DefineCriticalErrorAction(OnCriticalError);
-
-        if (IsSendOnly)
-        {
-            Session = new Session(Bus.CreateSendOnly(configuration));
-            return Task.FromResult(0);
-        }
-
-        configuration.PurgeOnStartup(!IsSeedingData && IsPurgingSupported);
-        Session = new Session(Bus.Create(configuration).Start());
-        return Task.FromResult(0);
-    }
-
-    BusConfiguration CreateConfiguration()
-    {
-        var configuration = new Configuration();
-        configuration.EndpointName(EndpointName);
-        configuration.EnableInstallers();
-
-        var scanableTypes = GetTypesToInclude();
-        configuration.TypesToScan(scanableTypes);
-
-        configuration.ApplyProfiles(this);
-
-        return configuration;
-    }
-
-    List<Type> GetTypesToInclude()
-    {
-        var location = System.IO.Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-        var asm = new NServiceBus.Hosting.Helpers.AssemblyScanner(location).GetScannableAssemblies();
-
-        var allTypes = (from a in asm.Assemblies
-                        from b in a.GetLoadableTypes()
-                        select b).ToList();
-
-        var allTypesToExclude = GetTypesToExclude(allTypes);
-
-        var finalInternalListToScan = allTypes.Except(allTypesToExclude);
-
-        return finalInternalListToScan.ToList();
-    }
-
-    void OnCriticalError(string errorMessage, Exception exception)
-    {
-        try
-        {
-            try
-            {
-                Log.Fatal("OnCriticalError", exception);
-                Session.CloseWithSuppress().ConfigureAwait(false).GetAwaiter().GetResult();
-            }
-            finally
-            {
-                NLog.LogManager.Shutdown();
-            }
-        }
-        finally
-        {
-            Environment.FailFast("NServiceBus critical error", exception);
-        }
-    }
-
-#else
-    async Task CreateOrPurgeAndDrainQueues()
-    {
-        var configuration = CreateConfiguration();
-        if (IsPurgingSupported) configuration.PurgeOnStartup(true);
-        ShortcutBehavior.Shortcut = true; // Required, as instance already receives messages before DrainMessages() is called!
-        var instance = await Endpoint.Start(configuration).ConfigureAwait(false);
-        await DrainMessages().ConfigureAwait(false);
-        await new Session(instance).CloseWithSuppress().ConfigureAwait(false);
-    }
-
-    async Task CreateSendOnlyEndpoint()
-    {
-        var configuration = CreateConfiguration();
-        configuration.SendOnly();
-        var instance = await Endpoint.Start(configuration).ConfigureAwait(false);
-        Session = new Session(instance);
-    }
-
-    async Task CreateEndpoint()
-    {
-        var configuration = CreateConfiguration();
-
-        var x = this as IConfigureUnicastBus;
-
-        if (x != null)
-        {
-            var routing = new RoutingSettings<FakeAbstractTransport>(configuration.GetSettings());
-
-            foreach (var m in x.GenerateMappings())
-            {
-                var isEvent = typeof(IEvent).IsAssignableFrom(m.MessageType);
-                if (isEvent) routing.RegisterPublisher(m.MessageType, m.Endpoint);
-                else routing.RouteToEndpoint(m.MessageType, m.Endpoint);
-
-            }
-        }
-
-        if (IsSendOnly)
-        {
-            configuration.SendOnly();
-            Session = new Session(await Endpoint.Start(configuration).ConfigureAwait(false));
-            return;
-        }
-
-        //configuration.PurgeOnStartup(!IsSeedingData && IsPurgingSupported);
-
-        var instance = Endpoint.Start(configuration).ConfigureAwait(false).GetAwaiter().GetResult();
-        Session = new Session(instance);
-    }
-
-    Configuration CreateConfiguration()
-    {
-        var configuration = new Configuration(EndpointName);
-        configuration.EnableInstallers();
-        configuration.AssemblyScanner().ExcludeTypes(GetTypesToExclude().ToArray());
-        configuration.ApplyProfiles(this);
-        configuration.DefineCriticalErrorAction(OnCriticalError);
-        return configuration;
-    }
-
-    IEnumerable<Type> GetTypesToExclude()
-    {
-        return GetTypesToExclude(Assembly.GetAssembly(this.GetType()).GetTypes());
-    }
-
-    async Task OnCriticalError(ICriticalErrorContext context)
-    {
-        try
-        {
-            try
-            {
-                Log.Fatal("OnCriticalError", context.Exception);
-                await context.Stop();
-            }
-            finally
-            {
-                NLog.LogManager.Shutdown();
-            }
-        }
-        finally
-        {
-            Environment.FailFast("NServiceBus critical error", context.Exception);
-        }
-    }
-
-#endif
 
     IEnumerable<Type> GetTypesToExclude(IEnumerable<Type> allTypes)
     {
